Casting to bigint preventing errors for larger xmin

+++ postgres/diag/sql/repl-xmin-from-activity.sql
@@ -1,1 +1,1 @@
-select now(),min(backend_xid::text::int), min(backend_xmin::text::int), max(backend_xid::text::int), max(backend_xmin::text::int) from pg_stat_activity;
+select now(),min(backend_xid::text::bigint), min(backend_xmin::text::bigint), max(backend_xid::text::bigint), max(backend_xmin::text::bigint) from pg_stat_activity;
